+++ lib/hubbado/upsert_version/controls.rb
@@ -1,5 +1,6 @@
 require_relative 'controls/id'
 require_relative 'controls/version'
 require_relative 'controls/attributes'
+require_relative 'controls/inserted'
 require_relative 'controls/unchanged'
-require_relative 'controls/upserted'
+require_relative 'controls/updated'